Arbitrary Missile Dice
Changed how missiles are handled to break the assumption that there's
always 2 dice per missile module which is no longer true in Second Dawn.
Also all 4 colors of missiles appear in Second Dawn.

Rearranged the UI a bit for the new missiles and general ergonomics.

Added a Hans Fugal copyright.